add FeatureMetaDataTweaker (not finished)

+++ java/src/org/antlr/codebuff/FeatureMetaData.java
@@ -3,7 +3,7 @@ package org.antlr.codebuff;
 public class FeatureMetaData {
 	public String[] abbrevHeaderRows;
 	public FeatureType type;
-	public int mismatchCost;
+	public double mismatchCost;
 
 	public FeatureMetaData(FeatureType type, String[] abbrevHeaderRows, int mismatchCost) {
 		this.abbrevHeaderRows = abbrevHeaderRows;
